Add internationalization support
This change adds i18n support so the script will work regardless of
whether the Audacity interface is in English or German.  Formerly,
the program failed when the Audacity language was set to English,
because it was trying to click a German "Exportieren" button that
was not present.  As of this change only English and German is supported
but a very basic framework to add other locales is in place.