Delegate file tagging to python
Interacting through the UI using apple scripts is inconsistent across
contexts and not reliable, so this change moves the tagging portion of
the program to follow the file export so it can be done purely in python
which is way more reliable, testable, fast and secure.

This change just skips through Audacity's metadata prompts and after the
files are exported, uses mutagen EasyID3 to set the tags and then renames
the mp3s themselves so the user doesn't need to bother naming the labels
in Audacity either.

A mechanism is in place to make sure that the number of files matches
the number of tracks on the discogs release and if it doesn't match, the
program will abort and give clear feedback about it.

If there are not enough labels, resulting in too few files, the program
will timeout after 20 seconds and tell the user.  The same is true with
too many labels (In this case the metadata prompts will never finish and
the python program will still time out after 20 seconds).

Additional security measures like ignoring files created before the
execution of the program are also in place, to minimize the possibility
of changing unrelated files that happen to be in the target directory.
Please note that old files can still be overwritten if they have the
exact same name that the program uses for the new files.

This also re-adds i18n support that was removed seemingly
unintentionally?